requirements.txt
ipykernel==6.25.1
jupyter_client==8.3.0
numpy==1.24.2
opencv_python==4.8.0.76
pytest==7.3.1
python-dotenv==1.0.0
setuptools==59.6.0
grpcio-tools==1.57.0
pytest-asyncio==0.21.0
grpc-google-iam-v1==0.12.6
matplotlib==3.7.2
pydantic==2.3.0
langchain==0.0.277
aiosqlite==0.19.0
rich==13.5.2
prompt_toolkit==3.0.39
aiofiles==23.2.1
click==8.1.7
discord.py==2.3.2
openai==0.28.0
